+++ CHANGELOG.md
@@ -26,3 +26,8 @@
   the slider in the direction of the click. Fixing not colored corners of slider 
   with active color in some cases. Correcting Timer declaring due to the sometimes 
   appearing exception.
+
+## [2.0.0] - 08.10.2021
+* Renamed bottomPadding to underSpacing, sliderPadding to sliderSpacing, bottomColor 
+* to underColor and bottomDecoration to underDecoration so as not to get confused 
+* with top/bottom and margin/padding. Some fixes in README.md